+++ src/exec/exec.c
@@ -3,64 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   exec.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lscheupl <lscheupl@student.42.fr>          +#+  +:+       +#+        */
+/*   By: leonel <leonel@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/06 17:41:00 by leonel            #+#    #+#             */
-/*   Updated: 2025/02/06 18:55:16 by lscheupl         ###   ########.fr       */
+/*   Updated: 2025/02/07 17:40:24 by leonel           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/minishell.h"
 #include <stdio.h>
 
-void	ft_execve(char *path, char **tab_arg, char **envp)
-{
-	execve(path, tab_arg, envp);
-	perror("execve");
-	exit(EXIT_FAILURE);
-}
 
-char	*quote_remover(char *str)
-{
-	int		i;
-	int		j;
-	char	*res;
-
-	i = 0;
-	j = 0;
-	res = malloc(sizeof(char) * (ft_strlen(str) + 1));
-	while (str[i])
-	{
-		if (str[i] == '\'')
-		{
-			i++;
-			while (str[i] != '\'')
-			{
-				res[j] = str[i];
-				i++;
-				j++;
-			}
-			i++;
-		}
-		if (str[i] == '\"')
-		{
-			i++;
-			while (str[i] != '\"')
-			{
-				res[j] = str[i];
-				i++;
-				j++;
-			}
-			i++;
-		}
-		res[j] = str[i];
-		j++;
-		i++;
-	}
-	res[j] = '\0';
-	free(str);
-	return (res);
-}
 
 void	redir_handler(int *redir, t_ms *ms, char **args)
 {
@@ -110,8 +63,8 @@ void	redir_handler(int *redir, t_ms *ms, char **args)
 		{
 			// printf("redir_hd\n");
 			fd = redir_hd(args[i + 1]);
-			// dup2(fd, ms->fd[0]);
-			// // close(fd);
+			if (fd == -1)
+				printf("error\n");  
 			redir[0] = fd;
 			spl_remove(args, i + 1);
 			spl_remove(args, i);
@@ -121,19 +74,6 @@ void	redir_handler(int *redir, t_ms *ms, char **args)
 	}
 }
 
-void redir_executions(int *redir, t_ms *ms)
-{
-	if (redir[0] != -1)
-	{
-		dup2(redir[0], ms->fd[0]);
-		// close(redir[0]);
-	}
-	if (redir[1] != -1)
-	{
-		dup2(redir[1], ms->fd[1]);
-		// close(redir[1]);
-	}
-}
 int exec_builtin(t_node_cmd **node, t_ms *ms)
 {
 	if ((*node)->bltin == E_EXIT)
@@ -169,62 +109,6 @@ int is_builtin(char *cmd)
 	else if (ft_strncmp(cmd, "env", 3) == 0)
 		return (E_ENV);
 	return (E_NOTBLTIN);
-}
-
-int	exec_cmd(char *input, t_ast *root, t_ms *ms)
-{
-	t_node_cmd	*node;
-	char		*path;
-	pid_t		pid;
-
-	node = &(root->cmd);
-	node->args = expand_expand(ft_convert_pos_to_string(input, node->start, node->end), ms);
-	node->redir[0] = -1;
-	node->redir[1] = -1;
-	redir_handler(node->redir, ms, node->args);
-	node->bltin = is_builtin(node->args[0]);
-	if (node->bltin != E_NOTBLTIN)
-	{
-		ms_close_fd(ms);
-		return(exec_builtin(&node, ms));
-	}
-	path = ft_find_path(node->expanded, ms, node->args);
-	if (path == NULL)
-	{
-		fprintf(stderr, "command not found: %s\n", node->args[0]);
-		// exit_exec(ms);
-		ms_close_fd(ms);
-		return (ms->status = 127);
-	}
-	pid = fork();
-	if (pid == -1)
-	{
-        perror("fork");
-        return (/*exit_fork(ms)*/0);
-    }
-	if (pid == 0)
-	{
-		redir_executions(node->redir, ms);
-		ms_close_fd(ms);
-		if (node->redir[0] != -1)
-			close(node->redir[0]);
-		if (node->redir[1] != -1)
-			close(node->redir[1]);
-		ft_execve(path, node->args, ms->envp);
-		exit_exec(ms);
-	}
-	if (node->redir[0] != -1)
-		close(node->redir[0]);
-	if (node->redir[1] != -1)
-		close(node->redir[1]);
-	close(ms->fd[0]);
-	close(ms->fd[1]);
-	waitpid(pid, &ms->status, 0);
-	free(path);
-	
-	if (WIFSIGNALED(ms->status))
-		return (128 + WTERMSIG(ms->status));
-	return(WEXITSTATUS(ms->status));
 }
 
 int	exec_pip(char *input, t_ast *root, t_ms *ms)
